HBASE-22324 loss a mass of data when the sequenceId of cells greater than Integer.Max

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/MemStoreCompactorSegmentsIterator.java
@@ -59,7 +59,7 @@ public class MemStoreCompactorSegmentsIterator extends MemStoreSegmentsIterator 
     // create the list of scanners to traverse over all the data
     // no dirty reads here as these are immutable segments
     int order = segments.size();
-    AbstractMemStore.addToScanners(segments, Integer.MAX_VALUE, order, scanners);
+    AbstractMemStore.addToScanners(segments, Long.MAX_VALUE, order, scanners);
     // build the scanner based on Query Matcher
     // reinitialize the compacting scanner for each instance of iterator
     compactingScanner = createScanner(store, scanners);
